% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_ode_excel}
\alias{download_ode_excel}
\title{Download Excel file from ODE}
\usage{
download_ode_excel(url, type, end_year)
}
\arguments{
\item{url}{URL to download}

\item{type}{"public" or "nonpub"}

\item{end_year}{School year end}
}
\value{
Data frame or NULL if download fails
}
\description{
Helper function to download and read an Excel file from ODE.
}
\keyword{internal}
