% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_enrollment.R
\name{filter_county}
\alias{filter_county}
\title{Filter enrollment data by county}
\usage{
filter_county(df, county_name)
}
\arguments{
\item{df}{Enrollment dataframe}

\item{county_name}{Name of Ohio county (case-insensitive)}
}
\value{
Filtered data frame
}
\description{
Convenience function to filter enrollment data to a specific Ohio county.
}
\examples{
\dontrun{
# Get Franklin County schools
franklin <- fetch_enr(2024) |> filter_county("Franklin")
}
}
