% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_assessment.R
\name{get_assessment_cache_path}
\alias{get_assessment_cache_path}
\title{Get cache path for assessment data}
\usage{
get_assessment_cache_path(end_year, cache_type)
}
\arguments{
\item{end_year}{School year end}

\item{cache_type}{Cache type identifier}
}
\value{
Path to cached file
}
\description{
Get cache path for assessment data
}
\keyword{internal}
