% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_available_years}
\alias{get_available_years}
\title{Get available years for enrollment data}
\usage{
get_available_years()
}
\value{
A list with three elements:
\describe{
\item{min_year}{The earliest available school year end (e.g., 2007 = 2006-07)}
\item{max_year}{The most recent available school year end (e.g., 2025 = 2024-25)}
\item{description}{A human-readable description of data availability}
}
}
\description{
Returns information about the range of school years for which enrollment
data is available from the Ohio Department of Education and Workforce.
}
\examples{
years <- get_available_years()
years$min_year
years$max_year
}
