% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{get_raw_enr_legacy}
\alias{get_raw_enr_legacy}
\title{Download legacy format ODEW enrollment data}
\usage{
get_raw_enr_legacy(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Raw data frame
}
\description{
Downloads enrollment data for years prior to 2015 when data formats
and locations were different.
}
\keyword{internal}
