% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{list_assessment_years}
\alias{list_assessment_years}
\title{List available assessment data years}
\usage{
list_assessment_years()
}
\value{
Integer vector of available years
}
\description{
Returns a vector of years for which assessment data is available
from the Ohio DOE Azure blob storage.
}
\details{
Note: 2020 data does not exist due to the COVID-19 pandemic.
Ohio cancelled state testing in spring 2020.
}
\examples{
list_assessment_years()
# Returns: 2018 2019 2021 2022 2023 2024 2025
}
