% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_assessment.R
\name{write_assessment_cache}
\alias{write_assessment_cache}
\title{Write assessment data to cache}
\usage{
write_assessment_cache(data, end_year, cache_type)
}
\arguments{
\item{data}{Data frame to cache}

\item{end_year}{School year end}

\item{cache_type}{Cache type identifier}
}
\description{
Write assessment data to cache
}
\keyword{internal}
