% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{list_enr_years}
\alias{list_enr_years}
\title{List available enrollment data years}
\usage{
list_enr_years()
}
\value{
Integer vector of available years
}
\description{
Returns a vector of years for which enrollment data is likely available
from the Ohio School Report Cards portal.
}
\examples{
\dontrun{
available_years <- list_enr_years()
}
}
